NavX yaw is now negated
This seemed to fix the angle doubling issue initially, but we're still
having issues with it so that didn't seem to be the only problem. For
whatever reason, the PID controller occasionally starts at a setpoint of
something other than 0 (it has been as big as -60), so we should
probably look in to that.