Odometry & turning should both be more accurate
In TurnToAngleProfiled, we've been zeroing the gyro yaw without
resetting the odometry, which could make it inaccurate. DriveSubsystem's
resetHeading method now handles that as well. We also now reset the
heading before calling ProfiledPIDController's constructor, which resets
the internal PID controller. I believe this will eliminate the issue of
the robot turning alternating amounts.

When updating the odometry, I also forgot to change the degree
measurements of the NavX to radians, but I fixed that in this commit.

One thing I realized was that if we negated the yaw output from the
NavX, we would have to do the same with the yaw velocity. It's negated
now when logged to the dashboard, but if we expose it later we have to
keep that in mind.