Lots of different changes
Increased feedforward constant for turning with Smart Motion because of
increase friction due to the carpet.

Increased the current limit for the intake arm to 15 amps, and Increased
its max power to full power.

Undo port forwarding because it might be causing robot code to cut out?

Added reversal for intake and indexer.

Intake now runs for all of TaxiThenShoot autonomous.

Also fixed RampThenShoot, as it was throwing a NullPointerException in
the past.

Intake lift motors were changed to coast mode from brake mode to avoid
shredding the gears (they're aluminum).