Again, a bunch of different competition changes
Added arbitrary feedforward power for turning using Smart Motion.

Secondary camera feed is now functional, but for some reason the
limelight feed doesn't work on the driver station.

Adjusted ClimbTimed timings again.

Started implementing speed switching in DefaultDrive but didn't finish.

IntakeLift is meant to be used when the intake lift motors are in brake
mode to reduce wear on the gears.

Shooter ramp time in RampThenShoot was increased to a maximum of 1
second.