More competition changes!
Added some delays to TaxiThenShoot auto, as well as added intake homing
as the first thing.

Adjusted climb timings.

Add an adjustable RPM offset to ShooterSubsystem.

Change the behavior for ShooterMath.calculateRPM so that if the
Limelight picks up a target outside of its calibrated range, it reverts
to the default RPM.