RampThenShoot should now rumble both controllers
We had some trouble getting the secondary to rumble, so we should
probably check if that's an issue with the controller or just how we
rumble it.

The compressor is now enabled when telop starts (maybe) because we were
having some issues with it not running, but that might have been due to
it being at pressure.