Trajectory following works* now!
The robot no longer spins out when it tries to follow a trajectory, but
we still have to test accuracy. The spinning out was due to us swapping
the left and right drive Spark Maxes, which was fixed in this commit.
This also means the right stick powers for turning no longer have to be
negated.

(post-rebase note: We just had the wrong kFF value on the spark maxes;
the velocities should have been in meters per second. As of this commit
it doesn't actually work, but this is fixed in the next few commits.)