Clean up & document climber-related code
A lot of unused code (mostly lidar stuff) was removed from
ClimberSubsystem, and some method names were changed to be more clear in
what they do.

Resetting the climber arms was also moved to testInit from the
ClimberSubsystem constructor, since the pistons don't actually move
until the robot is enabled. This change was meant to make that more
clear.