Transition to a hybrid drive motor control method
I was reading through the FRC server, and someone mentioned that they
used WPILib feedforward with Spark Max closed-loop velocity control on
their drive motors. I thought that would be interesting to try, since
it could alleviate the pain of tuning kFF on the Spark Maxes, as well as
allowing acceleration to be accounted for. This commmit changes both the
DriveStraightProfiled (now DriveMeters) and FollowTrajectory commands to
utilize this form of hybrid control. As of now it is untested, but I
will do that when I get a chance.