Fix an exception in TrajectoryCommandGroup
I didn't realize that modifying a list while iterating over it would
throw a ConcurrentModificationException, although it makes sense. I
modified the affected loop so it iterates over a ListIterator, which
does allow for the modification of a list as its iterated over.

Spotless formatting was also applied, since I haven't done that in a
bit.