Return to a TrapezoidProfileCommand for driving straight
I didn't realize it earlier but attempting to generate a straight
trajectory that spans a short distance seems to cause malformed spline
errors for some reason. Because we do PID control on the Spark Maxes
themselves, I think that just using a TrapezoidProfileCommand for
driving distances should work fine (although this should be tested).