Bring back static feedforward component
This seemed to increase the forward/backward distance accuracy a bit,
but at the cost of leading to left/right overshoots for Test Path.

I also changed Test Path to drive forward and backward along the same
path, which makes verifying accuracy easier. That way I only have to
measure a couple inches instead of the full path distance.