Reorganize drivetrain constants
Right now, the only inner class it has is TurnPID, although I'll
probably work on removing that eventually.

I also fixed TurnDegrees (I think), since I changed its feedforward to
be in terms of volts without changing the rest of the logic. That
transition is outside of the scope of this commit.