Rearrange example TrajectoryCommandGroup auto
The homing and lowering of the intake seemed inconsistent when it was
run as part of the TrajectoryCommandGroup. I realized that those things
should probably be run before the robot starts moving, so I changed the
declaration to reflect that.

I also removed a commented out section of AlignWithHub, since there was
no reason to keep it around.

As always, Spotless was applied for formatting.